Defer profile photo capture and upload code to lazy module (#2339)
Extracts camera, image-processing, and upload logic from profileService.ts
into a new profilePhotoService.ts that is only imported when the user
initiates a photo action on the Profile page, reducing cold-start bundle
weight. Updates Profile.tsx to dynamic-import profilePhotoService on demand.
